==> 01-Initial-Visualization-and-Cleaning.rmd <==
---
title: "Plant Data"
subtitle: "01 Initial visualization and cleaning"
author: "Kevin Schoelz"
date: "August 2019"
output:
  tufte::tufte_html:
    toc: true
    toc_depth: 2
---

We have an experiment where we are looking at the behavior of a certain species of plant with different gene combinations that have been knocked out (different plant ecotypes). In this experiment we inoculate all of the plants with a virus. The plants are then observed once a day for several days to see if/when virus symptoms become apparent. Our goal is to determine whether or not the different genotypes take a different amount of time in order to show viral symptoms. We will analyze this plant data in a series of notebooks. In this notebook, we are getting our data into a clean/tidy format and visualizing some of the raw data.

First, we set up the notebook, and load some useful libraries.

```{r setup, echo=TRUE, message = FALSE, warning = FALSE, results='hide'}

# Clear the environment
rm(list = ls())
#gc(reset = TRUE, verbose = FALSE)

# Notebook Settings
knitr::opts_chunk$set(echo = FALSE)

# Load Libraries
library(tidyverse)
library(tufte)
library(cowplot)
library(readxl)
library(lubridate)

#ggplot2 settings
theme_set(theme_cowplot())

```

# Importing the Data

The data in which we are interested has been stored in an excel file `Plant-Data-Messy-Format.xlsx` in the `tidyplants_cdf` sheet. We want to import this data into a dataframe so we can work with it. Along the way, we will relabel some column names so that they are easier for us to code with, ensure that the ecotypes are listed in the proper order, as well as cache a snapshot of the data in `tidyplants_cdf.csv`. At the end of this notebook, the full dataset has been printed.

```{r load_data, message=FALSE, echo=FALSE}

# Load data and cache a csv snapshot of the excel files

# Load cdf data
tidyplants_cdf <- read_excel("Plant-Data-Messy-Format.xlsx", sheet = "tidyplants_cdf")%>%
  rename(day = Day,
         ecotype = Ecotype,
         trial_name = `Trial Name`,
         trial_index = `Trial Index`,
         trial_date = `Trial Date`,
         total_infections = `Total Infections`,
         new_infections = `New Infections`,
         Ntot = `Total Plants`,
         Nobs = `Total Observed`,
         Ncens = `Total Censored`,
         ycens = `Censored Day`, 
         XIK = `XI-K`,
         XI2 = `XI-2`,
         VIIIA = `VIII-A`,
         VIIIB = `VIII-B`,
         XIB = `XI-B`,
         XIA = `XI-A`) %>%
  mutate(trial_date = as.Date(trial_date))

tidyplants_cdf$ecotype <- factor(tidyplants_cdf$ecotype, levels = c("Col", 
                                                                    "XI-K",
                                                                    "XI-2",
                                                                    "VIII-A",
                                                                    "VIII-B",
                                                                    "XI-B",
                                                                    "XI-A",
                                                                    "CHUP",
                                                                    "CHUP/XI-2",
                                                                    "CHUP/XI-K",
                                                                    "XI-K/XI-2",
                                                                    "CHUP/XI-K/XI-2"))

tidyplants_cdf %>% write_csv("tidyplants_cdf.csv")

head(tidyplants_cdf)

```

This dataset, as given, tracks what is happening for a given ecotype and trial, as a function of the data. It also keeps track of how many plants of an ecotype were in a given trial, and how many of the plants showed symptoms before the end of the trial.

Lets trial and visualize the results of the trial in a simple plot. Since these are the raw results, it will be a little difficult to interpret. But it can be nice to see. Plots will show number of plants that have shown symptoms as a function of time. Plots have also been broken up into trial number and ecotype. The full plot is shown below.


```{r cdf_vis, echo=FALSE}

cdf_plots <- ggplot(data = tidyplants_cdf, aes(x = day, y = total_infections)) +
  #geom_point(color = "DodgerBlue4") + 
  geom_step(color = "Black") +
  facet_grid(trial_index ~ ecotype) +
  scale_x_continuous(name = "Day",
                     limits = c(5,15))+
  ylab("Total Infections")

cdf_plots

ggsave("cdf_plots.pdf", plot = cdf_plots, height = 15, width = 15)
```

Since the number of trials and ecotypes is so large, it is a little difficult to see all of the trials. We can remedy this by breaking up the plot into three plots. The initial trials (1-4) are shown in the first plot. The second trials (5-9) are shown in the second plot and the third plot consists of trials 10-15.

```{r cdf_vis_2, echo + TRUE}

cdf_plots1 <- ggplot(data = tidyplants_cdf %>% filter(trial_index == 1 | trial_index == 2 | trial_index == 3 | trial_index == 4), aes(x = day, y = total_infections)) +
  geom_point(color = "DodgerBlue4") + 
  geom_step(color = "Black") +
  facet_grid(trial_index ~ ecotype) +
  xlab("Day") +
  ylab("Total Infections")

cdf_plots2 <- ggplot(data = tidyplants_cdf %>% filter(trial_index == 5 | trial_index == 6 | trial_index == 7 | trial_index == 8 | trial_index == 9), aes(x = day, y = total_infections)) +
  geom_point(color = "DodgerBlue4") + 
  geom_step(color = "Black") +
  facet_grid(trial_index ~ ecotype) +
  xlab("Day") +
  ylab("Total Infections")

cdf_plots3 <- ggplot(data = tidyplants_cdf %>% filter(trial_index == 10 | trial_index == 11 | trial_index == 12 | trial_index == 13 | trial_index == 14 | trial_index == 15), aes(x = day, y = total_infections)) +
  geom_point(color = "DodgerBlue4") + 
  geom_step(color = "Black") +
  facet_grid(trial_index ~ ecotype) +
  xlab("Day") +
  ylab("Total Infections")

cdf_plots1
cdf_plots2
cdf_plots3

```

# Transforming data so that we track each individual plant

One of the questions we will be interested in looking at is how many new infections we observe on a given day. This will be helpful for some of the data analysis we will be conducting later. We have extracted this data and stored it in the dataframe below.

```{r data_reshaping, echo=FALSE}

# Start by constructing plant data for uncensored plants

trial_name <- c()
trial_index <- c()
trial_date <- c()
day <- c()
ecotype <- c()
XIK <- c()
XI2 <- c()
VIIIA <- c()
VIIIB <- c()
XIB <- c()
XIA <- c()
CHUP <- c()


for(i in 1:nrow(tidyplants_cdf %>% filter(day != 0))){
  trial_name <- c(trial_name, rep(tidyplants_cdf$trial_name[i], each = tidyplants_cdf$new_infections[i]) )
  trial_index <- c(trial_index, rep(tidyplants_cdf$trial_index[i], each = tidyplants_cdf$new_infections[i]) )
  trial_date <- c(trial_date, rep(ymd(tidyplants_cdf$trial_date[i]), each = tidyplants_cdf$new_infections[i]) )
  day <- c(day, rep(tidyplants_cdf$day[i], each = tidyplants_cdf$new_infections[i]) )
  ecotype <- c(ecotype, rep(as.character(tidyplants_cdf$ecotype[i]), each = tidyplants_cdf$new_infections[i]) )
  XIK <- c(XIK, rep(tidyplants_cdf$XIK[i], each = tidyplants_cdf$new_infections[i]) )
  XI2 <- c(XI2, rep(tidyplants_cdf$XI2[i], each = tidyplants_cdf$new_infections[i]) )
  VIIIA <- c(VIIIA, rep(tidyplants_cdf$VIIIA[i], each = tidyplants_cdf$new_infections[i]) )
  VIIIB <- c(VIIIB, rep(tidyplants_cdf$VIIIB[i], each = tidyplants_cdf$new_infections[i]) )
  XIB <- c(XIB, rep(tidyplants_cdf$XIB[i], each = tidyplants_cdf$new_infections[i]) )
  XIA <- c(XIA, rep(tidyplants_cdf$XIA[i], each = tidyplants_cdf$new_infections[i]) )
  CHUP <- c(CHUP, rep(tidyplants_cdf$CHUP[i], each = tidyplants_cdf$new_infections[i]) )
}

df_censored <- data.frame(trial_name = trial_name,
                          trial_index = trial_index,
                          trial_date = trial_date,
                          day = day,
                          ecotype = ecotype,
                          XIK = XIK,
                          XI2 = XI2,
                          VIIIA = VIIIA,
                          VIIIB = VIIIB,
                          XIB = XIB,
                          XIA = XIA,
                          CHUP = CHUP) %>% 
  mutate(censored = 0)

df_censored$ecotype <- factor(df_censored$ecotype, levels = c("Col", 
                                                              "XI-K",
                                                              "XI-2",
                                                              "VIII-A",
                                                              "VIII-B",
                                                              "XI-B",
                                                              "XI-A",
                                                              "CHUP",
                                                              "CHUP/XI-2",
                                                              "CHUP/XI-K",
                                                              "XI-K/XI-2",
                                                              "CHUP/XI-K/XI-2"))

df_censored %>% write_csv("tidyplants.csv")

head(df_censored, 200)

```

```{r}

tp_hist <- ggplot(data = df_censored, aes(day)) +
  geom_histogram(color = "Black", fill = "DodgerBlue4", binwidth = 1) + 
  facet_grid(trial_date ~ ecotype)

tp_hist1 <- ggplot(data = df_censored %>% filter(trial_index == 1 | trial_index == 2 | trial_index == 3 | trial_index == 4), aes(day)) +
  geom_histogram(color = "Black", fill = "DodgerBlue4", binwidth = 1) + 
  facet_grid(trial_date ~ ecotype)

tp_hist2 <- ggplot(data = df_censored %>% filter(trial_index == 5 | trial_index == 6 | trial_index == 7 | trial_index == 8 | trial_index == 9), aes(day)) +
  geom_histogram(color = "Black", fill = "DodgerBlue4", binwidth = 1) + 
  facet_grid(trial_date ~ ecotype)

tp_hist3 <- ggplot(data = df_censored %>% filter(trial_index == 10 | trial_index == 11 | trial_index == 12 | trial_index == 13 | trial_index == 14 | trial_index == 15), aes(day)) +
  geom_histogram(color = "Black", fill = "DodgerBlue4", binwidth = 1) + 
  facet_grid(trial_date ~ ecotype)

tp_hist
tp_hist1
tp_hist2
tp_hist3

ggsave("hist_plots.pdf", plot = tp_hist, height = 15, width = 15)
```

Now that we have our data exported out of the excel file and into `tidyplants.csv` and `tidyplants_cdf.csv` we can begin our analysis of the dataset.




# Dataframes

## `tidyplants_cdf.csv`

```{r tidyplants_cdf}

knitr::kable(tidyplants_cdf)

```

## `tidyplants.csv`
