==> 02-Classical-Survival-Analysis.Rmd <==
---
title: "02 Kaplan-Meier Analysis"
author: "Kevin Schoelz"
date: "August 2019"
output:
  tufte::tufte_html:
    toc: true
    toc_depth: 2
---

We have an experiment where we are looking at the behavior of different ecotypes of a certain species of plant. In the experiment, the plant are inoculated by a virus. They are then observed once a day for several days to see when virus symptoms become apparent. There is reason to believe that maybe some of the ecotypes can slow the infection of the plant. In this notebook, we want to analyze the data by using the Kaplan-Meier estimator of the survival function.

Let's start by loading up some useful libraries.

```{r setup, echo=FALSE, message=FALSE}

# Clear the environment
rm(list = ls())
#gc(reset = TRUE, verbose = FALSE)

# Notebook Settings
knitr::opts_chunk$set(echo = FALSE)

library(tidyverse)
library(broom)
library(tufte)
library(cowplot)

#ggplot2 settings
theme_set(theme_cowplot())

```

The data we are interested in, has been stored in the file `tidyplants_cdf.csv`. This data looks at observations recorded on each day of a given trial, for a given plant ecotype. This data is ideally situated for calcuating the Kaplan-Meier estimator of the survival function. We load this data below.

```{r load-data, echo=FALSE}

# Load data about new.infections pdf from csv file
tidyplants_cdf <- read.csv("tidyplants_cdf.csv")

# Display tables to see if everything worked.
head(raw_df)

```

# Kaplan Meier Analysis

## Survival Models

In a survival analysis, we have some object, (people, plants, machine etc...) that can undergo some type of event. In a survival analysis, our goal is to estimate the time until the event happens. The canonical example would be to analyze mortality curves, i.e. to estimate time until death for people who undergo some type of treatment. But the mathematical framework is more general than that. We can use survival analysis to analyze time until a disease shows up (as we will do here), or even time until a given part fails on a machine. It is a very general and powerful framework. 
In survival analysis, we are trying to estimate a series of functions. We have the probability distribution function $f(y)$ which gives the probability of a new infection occuring on day $y$. There is the cumulative distribution function, $F(y)$ which gives the probability that the event has occured by day $y$. Finally we have the survival function $S(y) = 1 - F(y)$ which gives the probability of a given individual surviving until day $y$. Classically, the goal of a survival analysis is to estimate the survival function $S(y)$. We can estimate the survival function by looking at the proportion or surviving individuals in the population. For our data, this is not too complicated. For datasets where individuals leave the experiment early, or at different times, you need to more carefully consider your construction of the survival function.

# Kaplan Meier Analysis

With some slight changes to the `tidyplants_cdf.csv` file we can get the Kaplan-Meier estimator of the survival function.

```{r}

tidyplants_cdf <- tidyplants_cdf %>%
  mutate(#trial_date = as.Date(trial_date, format = "%d.%m.%Y"),
         km_survival = 1-total_infections/Ntot,
         healthy_pool = Ntot - total_infections,
         km_helpercol1 = new_infections/(healthy_pool + new_infections),
         km_stan_error = km_survival*sqrt(cumsum(km_helpercol1)),
         km_95conf = 1.96*km_stan_error,
         km_95upper = km_survival + km_95conf,
         km_95lower = km_survival - km_95conf)

head(tidyplants_cdf)

```


```{r}

treatments <- tidyplants_cdf %>%
  filter(ecotype != "Col") %>%
  rename(total_infections_treat = total_infections,
         Ntot_treat = Ntot,
         Nobs_treat = Nobs,
         Ncens_treat = Ncens,
         ycens_treat = ycens,
         new_infections_treat = new_infections,
         healthy_pool_treat = healthy_pool,
         km_survival_treat = km_survival,
         km_helpercol1_treat = km_helpercol1,
         km_stan_error_treat = km_stan_error,
         km_95conf_treat = km_95conf,
         km_95upper_treat = km_95upper,
         km_95lower_treat = km_95lower)

controls <- tidyplants_cdf %>%
  filter(ecotype == "Col") %>%
  select(trial_name,
         day,
         total_infections,
         Ntot,
         Nobs,
         Ncens,
         ycens,
         new_infections,
         healthy_pool,
         km_survival,
         km_helpercol1,
         km_stan_error,
         km_95conf,
         km_95upper,
         km_95lower) %>%
  rename(total_infections_ctrl = total_infections,
         Ntot_ctrl = Ntot,
         Nobs_ctrl = Nobs,
         Ncens_ctrl = Ncens,
         ycens_ctrl = ycens,
         new_infections_ctrl = new_infections,
         healthy_pool_ctrl = healthy_pool,
         km_survival_ctrl = km_survival,
         km_helpercol1_ctrl = km_helpercol1,
         km_stan_error_ctrl = km_stan_error,
         km_95conf_ctrl = km_95conf,
         km_95upper_ctrl = km_95upper,
         km_95lower_ctrl = km_95lower)
  
controls

```

```{r}

km_cdf <- left_join(treatments, controls, by = c("trial_name",
                                                 "day"))

km_cdf

km_plot1 <- ggplot(data = km_cdf %>% filter(trial_index == 1 | trial_index == 2 | trial_index == 3 | trial_index == 4)) +
  geom_step(aes(x = day, y = km_survival_treat), color = "Maroon", size = 1) +
  geom_step(aes(x = day, y = km_survival_ctrl), color = "DodgerBlue4", size = 1) +
  facet_grid(trial_name ~ ecotype) +
  xlim(6,13) +
  ylab("Survival Probability") +
  xlab("Day")

km_plot2 <- ggplot(data = km_cdf %>% filter(trial_index == 5 | trial_index == 6 | trial_index == 7 | trial_index == 8 | trial_index == 9)) + 
  geom_step(aes(x = day, y = km_survival_treat), color = "Maroon", size = 1) +
  geom_step(aes(x = day, y = km_survival_ctrl), color = "DodgerBlue4", size = 1) +
  facet_grid(trial_name ~ ecotype) +
  xlim(6,13) +
  ylab("Survival Probability") +
  xlab("Day")

km_plot3 <- ggplot(data = km_cdf %>% filter(trial_index == 10 | trial_index == 11 | trial_index == 12 | trial_index == 13 | trial_index == 14 | trial_index == 15)) + 
  geom_step(aes(x = day, y = km_survival_treat), color = "Maroon", size = 1) +
  geom_step(aes(x = day, y = km_survival_ctrl), color = "DodgerBlue4", size = 1) +
  facet_grid(trial_name ~ ecotype) +
  xlim(6,13) +
  ylab("Survival Probability") +
  xlab("Day")

km_plot1
km_plot2
km_plot3

```
